// src/organizers/dtos/patch.organizer.dto.generated.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file.
 *
 * =>  @see ./src/schemas/organizers/patchOrganizer.json
 *
 * and run "npm run schema-to-interface" to regenerate this file.
 */

export interface PatchOrganizer {
  name?: string;
  description?: string;
}